Make shr-models and bunyan peer dependencies
Due to the nature of the SHR tools, the entire chain must be guaranteed to be using the same instance of the shr-models library.  This has caused issues in the past and required every component of the chain to be updated and re-published whenever shr-models changed in any way.  A better approach to this is to declare shr-models as a peer dependency and rely on the end-user application to bring in the single instance of shr-models.

Bunyan is used in a similar way, and thus also declared a peer dependency.

In order to run the tests, however, these libararies need to be available, so they've also been added as dev dependencies.

This also required modifying some tests based on changes in the updated shr-test-helpers.